refactor: Изменить файловую структуру компонента Timer
